+++ __mocks__/@wdio/config.ts
@@ -7,31 +7,6 @@ import {
     validateConfig as validateConfigMock
 } from '../../packages/wdio-config/src/utils.js'
 
-class ConfigParserMock {
-    addService = vi.fn()
-    initialize = vi.fn()
-    addConfigFile = vi.fn()
-    merge = vi.fn()
-    autoCompile = vi.fn()
-
-    getConfig = vi.fn().mockReturnValue({
-        runnerEnv: {},
-        shard: { current: 1, total: 1 },
-        runner: 'local',
-        outputDir: './tempDir'
-    })
-
-    getCapabilities = vi.fn().mockReturnValue([{
-        browserName: 'chrome',
-        specs: ['./tests/test2.js']
-    }, {
-        browserName: 'firefox'
-    }])
-
-    getSpecs = vi.fn().mockReturnValue(['./tests/test1.js'])
-}
-
-export const ConfigParser = ConfigParserMock
 export const DEFAULT_CONFIGS = DEFAULT_CONFIGS_IMPORT as () => Omit<Options.Testrunner, 'capabilities'>
 export const isCloudCapability = vi.fn().mockImplementation(isCloudCapabilityMock)
 export const validateConfig = vi.fn().mockImplementation((defaults, config) => {
