+++ packages/devtools/src/cjs/index.ts
@@ -1,13 +1,10 @@
-import type { Options } from '@wdio/types'
-import type { Client, AttachOptions } from '../types'
-
 class Devtools {
     static async newSession (
-        options: Options.WebDriver,
+        options: any,
         modifier?: (...args: any[]) => any,
         userPrototype = {},
         customCommandWrapper?: (...args: any[]) => any
-    ): Promise<Client> {
+    ): Promise<any> {
         const WebDriver = (await import('../index.js')).default
         return WebDriver.newSession(options, modifier, userPrototype, customCommandWrapper)
     }
@@ -16,11 +13,11 @@ class Devtools {
      * allows user to attach to existing sessions
      */
     static async attachToSession (
-        options: AttachOptions,
+        options: any,
         modifier?: (...args: any[]) => any,
         userPrototype = {},
         commandWrapper?: (...args: any[]) => any
-    ): Promise<Client> {
+    ): Promise<any> {
         const Devtools = (await import('../index.js')).default
         return Devtools.attachToSession(options, modifier, userPrototype, commandWrapper)
     }
@@ -32,7 +29,7 @@ class Devtools {
      * @param   {Object} instance  the object we get from a new browser session.
      * @returns {string}           the new session id of the browser
      */
-    static async reloadSession (instance: Client) {
+    static async reloadSession (instance: any) {
         const Devtools = (await import('../index.js')).default
         return Devtools.reloadSession(instance)
     }
