+++ packages/eslint-plugin-wdio/tsconfig.cjs.json
@@ -2,6 +2,7 @@
   "extends": "./tsconfig.prod.json",
   "compilerOptions": {
     "module": "CommonJS",
+    "moduleResolution": "Node",
     "outDir": "./cjs",
     "declaration": false,
     "declarationMap": false
