+++ packages/wdio-cli/src/launcher.ts
@@ -3,7 +3,8 @@ import path from 'node:path'
 import exitHook from 'async-exit-hook'
 
 import logger from '@wdio/logger'
-import { ConfigParser, validateConfig } from '@wdio/config'
+import { validateConfig } from '@wdio/config'
+import { ConfigParser } from '@wdio/config/node'
 import { initialisePlugin, initialiseLauncherService, sleep } from '@wdio/utils'
 import { setupDriver, setupBrowser } from '@wdio/utils/node'
 import type { Options, Capabilities, Services } from '@wdio/types'
