+++ packages/wdio-cli/src/utils.ts
@@ -16,7 +16,7 @@ import { $ } from 'execa'
 import { readPackageUp } from 'read-pkg-up'
 import { resolve } from 'import-meta-resolve'
 import { SevereServiceError } from 'webdriverio'
-import { ConfigParser } from '@wdio/config'
+import { ConfigParser } from '@wdio/config/node'
 import { CAPABILITY_KEYS } from '@wdio/protocols'
 import type { Capabilities, Options, Services } from '@wdio/types'
