+++ packages/wdio-cli/tests/watcher.test.ts
@@ -15,7 +15,7 @@ vi.mock('@wdio/logger', () => import(path.join(process.cwd(), '__mocks__', '@wdi
 vi.mock('@wdio/config', () => import(path.join(process.cwd(), '__mocks__', '@wdio/config')))
 vi.mock('@wdio/utils', () => import(path.join(process.cwd(), '__mocks__', '@wdio/utils')))
 vi.mock('../src/launcher', async () => {
-    const { ConfigParser } = await import('@wdio/config')
+    const { ConfigParser } = await import('@wdio/config/node')
 
     interface LauncherMockRunCommandArguments extends Omit<RunCommandArguments, 'configPath'> {
         isMultiremote?: boolean;
