+++ packages/wdio-config/package.json
@@ -8,7 +8,14 @@
   "type": "module",
   "types": "./build/index.d.ts",
   "exports": {
-    ".": "./build/index.js",
+    ".": {
+      "import": "./build/index.js",
+      "types": "./build/index.d.ts"
+    },
+    "./node": {
+      "import": "./build/node/index.js",
+      "types": "./build/node/index.d.ts"
+    },
     "./package.json": "./package.json"
   },
   "typeScriptVersion": "3.8.3",
