+++ packages/wdio-config/src/node/RequireLibrary.ts
