+++ packages/wdio-config/src/node/index.ts
@@ -1,0 +1,3 @@
+import ConfigParser from './ConfigParser.js'
+
+export { ConfigParser }
