+++ packages/wdio-config/tests/FileSystemPathService.test.ts
@@ -4,7 +4,7 @@ import fs from 'node:fs'
 import { sync as globSync } from 'glob'
 import { vi, describe, it, expect, afterEach } from 'vitest'
 
-import FileSystemPathService from '../src/lib/FileSystemPathService.js'
+import FileSystemPathService from '../src/node/FileSystemPathService.js'
 
 const INDEX_PATH = path.resolve(__dirname, '..', 'src', 'index.ts')
