+++ packages/wdio-config/tests/RequireLibrary.test.ts
@@ -1,4 +1,4 @@
-import RequireLibrary from '../src/lib/RequireLibrary.js'
+import RequireLibrary from '../src/node/RequireLibrary.js'
 import { vi, describe, it, expect } from 'vitest'
 
 vi.mock('ts-node', () => ({ default: 'mock module' }))
