+++ packages/wdio-cucumber-framework/src/index.ts
@@ -1,47 +1,22 @@
 import url from 'node:url'
 import path from 'node:path'
 import fs from 'node:fs'
+import { readdir, readFile } from 'node:fs/promises'
 import { createRequire } from 'node:module'
 import { EventEmitter } from 'node:events'
 import { Writable } from 'node:stream'
-import logger from '@wdio/logger'
+
+import got from 'got'
 import isGlob from 'is-glob'
 import { sync as globSync } from 'glob'
+
+import logger from '@wdio/logger'
 import { executeHooksWithArgs } from '@wdio/utils'
-import got from 'got'
-import { readdir, readFile } from 'node:fs/promises'
-
-import * as Cucumber from '@cucumber/cucumber'
-import Gherkin from '@cucumber/gherkin'
-import { IdGenerator } from '@cucumber/messages'
-
-import { DEFAULT_OPTS } from './constants.js'
-import { generateSkipTagsFromCapabilities } from './utils.js'
-
-import type {
-    CucumberOptions,
-    HookFunctionExtension as HookFunctionExtensionImport,
-    StepDefinitionOptions
-} from './types.js'
-import type { Feature, GherkinDocument } from '@cucumber/messages'
-import type { ITestCaseHookParameter } from '@cucumber/cucumber'
 import type { Capabilities, Options, Frameworks } from '@wdio/types'
 
-import type {
-    IRunEnvironment } from '@cucumber/cucumber/api'
+import { } from '@cucumber/cucumber'
 import {
-    loadConfiguration,
-    loadSources,
-    runCucumber
-} from '@cucumber/cucumber/api'
-
-import { DataTable, World, Status } from '@cucumber/cucumber'
-
-export const FILE_PROTOCOL = 'file://'
-
-const log = logger('@wdio/cucumber-framework')
-
-const {
+    default as Cucumber,
     After,
     AfterAll,
     AfterStep,
@@ -59,13 +34,31 @@ const {
     setWorldConstructor,
     defineParameterType,
     defineStep,
-} = Cucumber
+
+    DataTable,
+    World,
+    Status
+} from '@cucumber/cucumber'
+import Gherkin from '@cucumber/gherkin'
+import { IdGenerator } from '@cucumber/messages'
+import type { Feature, GherkinDocument } from '@cucumber/messages'
+import type { ITestCaseHookParameter } from '@cucumber/cucumber'
+import type { IRunEnvironment } from '@cucumber/cucumber/api'
+import { loadConfiguration, loadSources, runCucumber } from '@cucumber/cucumber/api'
+
+import { DEFAULT_OPTS } from './constants.js'
+import { generateSkipTagsFromCapabilities } from './utils.js'
+import type {
+    CucumberOptions,
+    HookFunctionExtension as HookFunctionExtensionImport,
+    StepDefinitionOptions
+} from './types.js'
+
+export const FILE_PROTOCOL = 'file://'
 
 const uuidFn = IdGenerator.uuid()
-
+const log = logger('@wdio/cucumber-framework')
 const require = createRequire(import.meta.url)
-
-const { incrementing } = IdGenerator
 
 function getResultObject(
     world: ITestCaseHookParameter
@@ -81,7 +74,7 @@ function getResultObject(
 
 class CucumberAdapter {
     private _cwd = process.cwd()
-    private _newId = incrementing()
+    private _newId = IdGenerator.incrementing()
     private _cucumberOpts: Required<CucumberOptions>
 
     private _hasTests = true
