+++ packages/wdio-sauce-service/src/service.ts
@@ -1,9 +1,12 @@
 import fs from 'node:fs/promises'
 import path from 'node:path'
 
+import {
+    default as SauceLabs,
+    type SauceLabsOptions,
+    type Job
+} from 'saucelabs'
 import logger from '@wdio/logger'
-import SauceLabs from 'saucelabs'
-import type { SauceLabsOptions, Job } from 'saucelabs'
 import type { Services, Capabilities, Options, Frameworks } from '@wdio/types'
 
 import { isRDC, ansiRegex } from './utils.js'
@@ -22,7 +25,7 @@ export default class SauceService implements Services.ServiceInstance {
     private _isJobNameSet = false
 
     private _options: SauceServiceConfig
-    private _api: SauceLabs
+    private _api: SauceLabs.default
     private _browser?: WebdriverIO.Browser | WebdriverIO.MultiRemoteBrowser
     private _isRDC?: boolean
     private _suiteTitle?: string
@@ -34,7 +37,6 @@ export default class SauceService implements Services.ServiceInstance {
         private _config: Options.Testrunner
     ) {
         this._options = { ...DEFAULT_OPTIONS, ...options }
-        // @ts-expect-error https://github.com/saucelabs/node-saucelabs/issues/153
         this._api = new SauceLabs.default(this._config as unknown as SauceLabsOptions)
         this._maxErrorStackLength = this._options.maxErrorStackLength || this._maxErrorStackLength
     }
