+++ packages/wdio-smoke-test-cjs-service/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "../../tsconfig",
   "compilerOptions": {
     "module": "CommonJS",
+    "moduleResolution": "Node",
     "esModuleInterop": true,
     "baseUrl": ".",
     "outDir": "./build",
