+++ packages/webdriver/src/cjs/index.ts
@@ -1,6 +1,3 @@
-import type { Options } from '@wdio/types'
-import type { Client, AttachOptions } from '../types.js'
-
 function command (method: string, encodeUri: string, commandInfo: any, doubleEncodeVariables = false) {
     return async function protocolCommand(this: unknown, ...args: unknown[]) {
         const commandESM = await import('../command.js')
@@ -10,11 +7,11 @@ function command (method: string, encodeUri: string, commandInfo: any, doubleEnc
 
 class WebDriver {
     static async newSession(
-        options: Options.WebDriver,
+        options: any,
         modifier?: (...args: any[]) => any,
         userPrototype = {},
         customCommandWrapper?: (...args: any[]) => any
-    ): Promise<Client> {
+    ): Promise<any> {
         const WebDriver = (await import('../index.js')).default
         return WebDriver.newSession(options, modifier, userPrototype, customCommandWrapper)
     }
@@ -23,11 +20,11 @@ class WebDriver {
      * allows user to attach to existing sessions
      */
     static async attachToSession (
-        options?: AttachOptions,
+        options?: any,
         modifier?: (...args: any[]) => any,
         userPrototype = {},
         commandWrapper?: (...args: any[]) => any
-    ): Promise<Client> {
+    ): Promise<any> {
         const WebDriver = (await import('../index.js')).default
         return WebDriver.attachToSession(options, modifier, userPrototype, commandWrapper)
     }
@@ -39,7 +36,7 @@ class WebDriver {
      * @param   {object} instance  the object we get from a new browser session.
      * @returns {string}           the new session id of the browser
      */
-    static async reloadSession (instance: Client) {
+    static async reloadSession (instance: any) {
         const WebDriver = (await import('../index.js')).default
         return WebDriver.reloadSession(instance)
     }
